Refactor expression hir lowering into cc file
